refactor: change from http put requests to patch requests to minimize…

+++ server/src/main/java/com/unihub/app/dto/community/content/response/EventResponseDto.java
@@ -22,7 +22,7 @@ public class EventResponseDto{
         private String description;
         private String locationDetails;
         private OffsetDateTime startTime;
-        private Double durationHours;
+        private Float durationHours;
         private EventLocation location;
         private String courseSlug;
         private String courseName;
